requirements.txt
gunicorn
flask
torch
Pillow
werkzeug
utils
numpy
matplotlib
pyplot
torchvision
models
detection